fix: speed up the startup sequence and add splash screen during loading
