+++ splash.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <link rel='stylesheet' href='static/css/base.css'>
+    <meta name="viewport" content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no"/>
+    <meta name='apple-mobile-web-app-capable' content='yes'/>
+    <meta name='apple-mobile-web-app-status-bar-style' content='black-translucent'/>
+  </head>
+  <body>
+  <div id="welcome" class='mapeo-container'>
+    <div id="screen-1">
+      <h1>
+        Cargando...
+      </h1>
+      <img src='static/mapeo_256x256.png' />
+    </div>
+  </div>
+  </body>
+</html>
