+++ src/renderer/components/Loading.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default class Loading extends React.Component {
+  render () {
+    return <div>Loading...</div>
+  }
+}
